fix(rules): only ship route templates whose .srs files actually exist
The Rules.vue template dropdown was hardcoded with three URLs that
404 on github.com/SagerNet/sing-geosite/rule-set:

  geosite-malware.srs
  geosite-phishing.srs
  geosite-cryptominers.srs

sing-geosite is derived from v2fly geosite and does NOT carry security
categories. Once any of these were added to a panel and saved, the next
sing-box start tried to fetch them, hit 404, cancelled the parallel
downloads of the other rule-sets, and the whole core refused to start:

  ERROR start sing-box err: initialize rule-set[2]: ... 404 Not Found
                          | initialize rule-set[1]: ... context canceled
                          | initialize rule-set[4]: ... context canceled

Replace the broken templates with rule-sets that actually exist on the
rule-set branch (verified 2026-05):
  - block-tracker  → geosite-category-public-tracker.srs
  - block-porn     → geosite-category-porn.srs
The good ones (block-ads / cn-direct / private-direct) keep the same
URLs. Recommended bundle drops the broken pair.

Also fix a long-standing logging bug exposed by the resulting cooldown
loop:

  service/config.go:99
    logger.Info("start core cooldown ", startCooldown/time.Second, ...)
                                       ^^ time.Duration / time.Duration
                                          is still a time.Duration; its
                                          String() printed "15ns".

Cast to int64 so the message reads "15 seconds" again.

Recovery for affected installs: open the panel → Rules → delete the
tmpl-malware / tmpl-phishing / tmpl-cryptominers rule-sets and the
rules referencing them, then save and restart sing-box.

+++ service/config.go
@@ -96,7 +96,9 @@ func (s *ConfigService) StartCore() error {
 		return nil
 	}
 	if time.Since(lastStartFailTime) < startCooldown {
-		logger.Info("start core cooldown ", startCooldown/time.Second, " seconds")
+		// startCooldown / time.Second 仍是 time.Duration 类型,String() 输出
+		// 它的纳秒数(如 "15ns")。强转 int64 才得到字面量秒数。
+		logger.Info("start core cooldown ", int64(startCooldown/time.Second), " seconds")
 		startCoreMu.Unlock()
 		return nil
 	}
